Extracting ICU/IConv from core engine
git-svn-id: https://svn.code.sf.net/p/quex/code/trunk@4746 db9deaaf-66c7-40b0-811a-ad7aae109365

+++ TEST/2896732/uXa.cpp
@@ -178,7 +178,7 @@ namespace quex {
 		} else {
 			size_t needCount = (peekIndex + 1) - queueSize;
 
-			for(int index = 0x00;index < needCount;index++) {
+			for(int index = 0x00;index < (int)needCount;index++) {
                 Token*  token_p = 0x0;
                 // Receive from lexer...
                 lexer.receive(&token_p);
